chore: add typing
