+++ connectors/torus.ts
@@ -1,8 +1,9 @@
 const get = () => import(/* webpackChunkName: "torus" */ '@toruslabs/torus-embed');
 import LockConnector from '../src/connector';
+import { EIP1193Provider } from '../src/types';
 
 export default class Connector extends LockConnector {
-  async connect() {
+  async connect(): Promise<EIP1193Provider | undefined> {
     let provider;
     try {
       const Torus = (await get()).default;
