+++ src/types.ts
@@ -1,0 +1,39 @@
+// Interface for Ethereum providers based on the EIP-1193 standard.
+export interface EIP1193Provider {
+  isStatus?: boolean;
+  host?: string;
+  path?: string;
+  sendAsync?: (request: { method: string, params?: Array<unknown> }, callback: (error: Error | null, response: unknown) => void) => void;
+  send?: (request: { method: string, params?: Array<unknown> }, callback: (error: Error | null, response: unknown) => void) => void;
+  request: (request: { method: string, params?: Array<unknown> }) => Promise<unknown>;
+}
+
+declare global {
+  interface WindowEventMap {
+    "eip6963:announceProvider": EIP6963AnnounceProviderEvent;
+  }
+}
+
+export class EIP6963RequestProviderEvent extends Event {
+  constructor() {
+    super("eip6963:requestProvider");
+  }
+}
+
+export interface EIP6963AnnounceProviderEvent extends Event {
+  type: "eip6963:announceProvider";
+  detail: EIP6963ProviderDetail;
+}
+
+export interface EIP6963ProviderDetail {
+  info: EIP6963ProviderInfo;
+  provider: EIP1193Provider;
+}
+
+export interface EIP6963ProviderInfo {
+  uuid: string;
+  name: string;
+  icon: string;
+  rdns: string;
+}
+
